refactor: AuthenticationHandler와 CurrentSecurityUserGetter를 하나의 클래스 AuthenticationRepository로 리팩터링
