--- src/main/java/com/map/gaja/global/authentication/AuthenticationHandler.java
@@ -1,24 +1,0 @@
-package com.map.gaja.global.authentication;
-
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContext;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.stereotype.Component;
-
-@Component
-public class AuthenticationHandler {
-    public void saveContext(Long userId, String email, String authority) {
-        UserDetails userDetails = new PrincipalDetails(userId, email, authority);
-        SecurityContext context = SecurityContextHolder.createEmptyContext();
-        context.setAuthentication(new UsernamePasswordAuthenticationToken(userDetails, null, userDetails.getAuthorities()));
-        SecurityContextHolder.setContext(context);
-    }
-
-    public String getEmail() {
-        PrincipalDetails principalDetails = (PrincipalDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        return principalDetails.getName();
-    }
-
-}
