+++ src/main/java/com/map/gaja/global/exception/GlobalExceptionHandler.java
@@ -1,12 +1,11 @@
 package com.map.gaja.global.exception;
 
-import com.map.gaja.global.authentication.AuthenticationHandler;
+import com.map.gaja.global.authentication.AuthenticationRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.http.converter.HttpMessageNotReadableException;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.validation.BindException;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -19,14 +18,14 @@ import java.util.List;
 @Slf4j
 @RequiredArgsConstructor
 public class GlobalExceptionHandler {
-    private final AuthenticationHandler authenticationHandler;
+    private final AuthenticationRepository authenticationRepository;
 
     /**
      * API 비즈니스 관련 예외 공동 처리
      */
     @ExceptionHandler(BusinessException.class)
     public ResponseEntity<ExceptionDto> handleBusinessError(BusinessException e) {
-        log.info("{}: {} => {}", authenticationHandler.getEmail(), e.getStatus(), e.getMessage());
+        log.info("{}: {} => {}", authenticationRepository.getEmail(), e.getStatus(), e.getMessage());
         return ResponseEntity
                 .status(e.getStatus())
                 .body(new ExceptionDto(e.getMessage()));
@@ -37,7 +36,7 @@ public class GlobalExceptionHandler {
      */
     @ExceptionHandler(WebException.class)
     public ResponseEntity<Object> handleWebError(WebException e) {
-        log.info("{}: {} => {}", authenticationHandler.getEmail(), e.getStatus(), e.getMessage());
+        log.info("{}: {} => {}", authenticationRepository.getEmail(), e.getStatus(), e.getMessage());
         return ResponseEntity
                 .status(e.getStatus())
                 .body(e.getBody());
@@ -45,7 +44,7 @@ public class GlobalExceptionHandler {
 
     @ExceptionHandler(Exception.class)
     public ResponseEntity<ExceptionDto> allHandle(Exception e) {
-        log.error("{}: {}", authenticationHandler.getEmail(), e);
+        log.error("{}: {}", authenticationRepository.getEmail(), e);
         return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
                 .body(new ExceptionDto("서버 에러"));
     }
@@ -58,7 +57,7 @@ public class GlobalExceptionHandler {
      */
     @ExceptionHandler({MethodArgumentNotValidException.class, BindException.class})
     public ResponseEntity<CommonErrorResponse> validationErrorHandle(BindException e) {
-        log.info("{}: {} => {}", authenticationHandler.getEmail(), e.getAllErrors().toString(), e.getMessage());
+        log.info("{}: {} => {}", authenticationRepository.getEmail(), e.getAllErrors().toString(), e.getMessage());
         List<ValidationErrorResponse> body = new ArrayList<>();
         e.getAllErrors().stream().forEach(
                 error -> body.add(
@@ -77,7 +76,7 @@ public class GlobalExceptionHandler {
     @ExceptionHandler
     public ResponseEntity<CommonErrorResponse> Handle(HttpMessageNotReadableException e) {
         CommonErrorResponse body = new CommonErrorResponse("Type-Mismatch", "Type-Mismatch");
-        log.info("{}: {}", authenticationHandler.getEmail(), e.getMessage());
+        log.info("{}: {}", authenticationRepository.getEmail(), e.getMessage());
         return new ResponseEntity(body, HttpStatus.BAD_REQUEST);
     }
 }
